// project-evaluations/scripts/sync-benchmarks/enums.ts
export enum Protocol {
  BLANKSQUARE = "blanksquare",
  CURVY = "curvy",
  FLUIDKEY = "fluidkey",
  HINKAL = "hinkal",
  HOUDINISWAP = "houdiniswap",
  INTMAX = "intmax",
  MONERO = "monero",
  PRIVACY_POOLS = "privacy-pools",
  RAILGUN = "railgun",
  REDACT = "redact",
  TORNADO_CASH = "tornado-cash",
  VEIL_CASH = "veil-cash",
  WORM = "worm",
  ZERC20 = "zerc20",
}

export enum Network {
  MAINNET = "mainnet",
  ARBITRUM = "arbitrum",
  SEPOLIA = "sepolia",
  BASE = "base",
  SCROLL = "scroll",
  STARKNET = "starknet",
}